refactor: move email validation logic into separate module with tests
